requirements.txt
opencv-python
numpy
matplotlib
Pillow
tflite-runtime